+++ CHANGELOG.md
@@ -1,3 +1,6 @@
+## Unreleased
+[no unreleased changes yet]
+
 ## v4.4.1 (2025-02-10)
 - Remove `RSpec/DescribeClass` custom config.