+++ CHANGELOG.md
@@ -1,4 +1,4 @@
-## Unreleased
+## v4.5.0 (2025-02-11)
 - Add custom `RungerStyle/MultilineHashValueIndentation` cop.
 - Add custom `RungerStyle/FirstArgumentIndentation` cop.